Move chaos utils
Move chaos utils for use in other modules
- create_vm_with_nginx_service
- verify_vm_service_reachable
- is_http_ok
